Scripting: undo refactor
