+++ modules/lang-painless/src/doc/java/org/elasticsearch/painless/ContextGeneratorCommon.java
@@ -16,7 +16,6 @@ import org.elasticsearch.painless.action.PainlessContextClassInfo;
 import org.elasticsearch.painless.action.PainlessContextInfo;
 import org.elasticsearch.painless.action.PainlessContextInstanceBindingInfo;
 import org.elasticsearch.painless.action.PainlessContextMethodInfo;
-import org.elasticsearch.painless.javadoc.JavadocExtractor;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
