+++ modules/lang-painless/src/doc/java/org/elasticsearch/painless/JavadocExtractor.java
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.painless.javadoc;
+package org.elasticsearch.painless;
 
 import com.github.javaparser.StaticJavaParser;
 import com.github.javaparser.ast.CompilationUnit;
